refactor: update product service import paths for consistency
