+++ src/app/api/products/[product_id]/route.ts
@@ -1,7 +1,7 @@
 import { NextRequest, NextResponse } from "next/server";
 
 import { serializeProduct } from "@/lib/utils";
-import { productService } from "@/services/product";
+import { productService } from "@/services/product/product.service";
 import { createErrorResponse, createSuccessResponse } from "@/types";
 import { SerializedProductDetail } from "@/types/product.types";
