+++ src/components/shops/individual-product.tsx
@@ -3,7 +3,7 @@ import { notFound } from "next/navigation";
 
 import { serializeProduct } from "@/lib/utils";
 import reviewRepository from "@/repositories/reviews.repository";
-import { productService } from "@/services/product";
+import { productService } from "@/services/product/product.service";
 import { ReviewWithUser } from "@/types/review.type";
 
 import ProductActions from "./product-actions";
